+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -474,9 +474,11 @@
     <Compile Include="Routing\DynamicRouting.cs" />
     <Compile Include="Routing\DynamicRoutingDefinition.cs" />
     <Compile Include="Routing\RoutingExtensions.cs" />
+    <Compile Include="Routing\FileBasedRoundRobinDynamicRouting.cs" />
+    <Compile Include="Routing\FileBasedDynamicRouting.cs" />
     <Compile Include="Routing\IRouterDistributor.cs" />
-    <Compile Include="Routing\RoutingDistributor.cs" />
-    <Compile Include="Routing\RoutingDistributorDefinition.cs" />
+    <Compile Include="Routing\DynamicRouting.cs" />
+    <Compile Include="Routing\DynamicRoutingDefinition.cs" />
     <Compile Include="Routing\RoutingExtentions.cs" />
     <Compile Include="Unicast\Transport\Config\TransportConfiguration.cs" />
     <Compile Include="Unicast\Transport\Config\TransportExtensions.cs" />
