+++ src/NServiceBus.Core/Routing/FileBasedRoundRobinDynamicRouting.cs
@@ -1,0 +1,18 @@
+namespace NServiceBus.Features
+{
+    using System;
+    using NServiceBus.Routing;
+
+    class FileBasedRoundRobinDynamicRouting : Feature
+    {
+        public FileBasedRoundRobinDynamicRouting()
+        {
+            DependsOn<DynamicRouting>();
+        }
+
+        protected internal override void Setup(FeatureConfigurationContext context)
+        {
+            context.Container.RegisterSingleton(new FileBasedRoundRobinDynamicRoutingImpl(context.Settings.Get<string>("FileBasedRouting.BasePath"), TimeSpan.FromSeconds(5)));
+        }
+    }
+}
