--- src/NServiceBus.Core/Routing/RoutingDistributorDefinition.cs
@@ -1,15 +1,0 @@
-namespace NServiceBus.Routing
-{
-    using System;
-
-    /// <summary>
-    /// Implemented by routing distributors to provide their capabilities
-    /// </summary>
-    public abstract class RoutingDistributorDefinition
-    {
-        /// <summary>
-        /// The feature to enable when this routing distributor is selected
-        /// </summary>
-        protected internal abstract Type ProvidedByFeature();
-    }
-}
